requirements.txt
streamlit==1.32.0
pandas==2.2.1
numpy==1.26.4
scikit-learn==1.4.1.post1
shap==0.44.1
evidently==0.3.3
psycopg2-binary==2.9.9
python-dotenv==1.0.1
plotly==5.18.0
joblib==1.3.2
pydantic==1.10.13
matplotlib==3.8.3
